requirements.txt
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
SQLModel
faker
numpy